+++ backend/KAZABUILD.Application/DTOs/Components/Components/MotherboardComponent/UpdateMotherboardComponentDto.cs
@@ -19,6 +19,6 @@ namespace KAZABUILD.Application.DTOs.Components.Components.MotherboardComponent
         public string? RAMType { get; set; }
 
         [Range(0, 15, ErrorMessage = "RAM Slots Amount must be between 1 and 15")]
-        public decimal? RAMSlotsAmount { get; set; }
+        public int? RAMSlotsAmount { get; set; }
     }
 }
